Galaxy: Fix --galaxy-install and --galaxy-show-builds for Linux games
The changes made in 7653a52631697138b3099f49e6267f95249bc100 broke --galaxy-install and --galaxy-show-builds for Linux games
Product builds json is empty for Linux games but Downloader::sortGalaxyProductBuilds didn't check for empty json and proceeded to add empty items array to it.
This broke a check later in the code because the json wasn't empty anymore.